+++ packages/create-poyo-app/CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project are documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+
+## [0.4.1] - 2026-09-08
+
+### Fixed
+
+- Restores `.gitignore` from `.gitignore.template` at root and client package root during scaffolding
+- Auto-renames `Poyo.Server` React build asset paths in `.gitignore` to use the dynamic project name
 ## [0.4.0] - 2026-08-20
 
 ### Added